+++ griptape/common/__init__.py
@@ -1,8 +1,14 @@
+from .actions.base_action import BaseAction
+from .actions.tool_action import ToolAction
+
 from .prompt_stack.contents.base_message_content import BaseMessageContent
 from .prompt_stack.contents.base_delta_message_content import BaseDeltaMessageContent
 from .prompt_stack.contents.text_delta_message_content import TextDeltaMessageContent
 from .prompt_stack.contents.text_message_content import TextMessageContent
 from .prompt_stack.contents.image_message_content import ImageMessageContent
+from .prompt_stack.contents.action_call_delta_message_content import ActionCallDeltaMessageContent
+from .prompt_stack.contents.action_call_message_content import ActionCallMessageContent
+from .prompt_stack.contents.action_result_message_content import ActionResultMessageContent
 
 from .prompt_stack.messages.base_message import BaseMessage
 from .prompt_stack.messages.delta_message import DeltaMessage
@@ -11,6 +17,7 @@ from .prompt_stack.messages.message import Message
 from .prompt_stack.prompt_stack import PromptStack
 
 from .reference import Reference
+
 
 __all__ = [
     "BaseMessage",
@@ -21,6 +28,11 @@ __all__ = [
     "TextDeltaMessageContent",
     "TextMessageContent",
     "ImageMessageContent",
+    "ActionCallDeltaMessageContent",
+    "ActionCallMessageContent",
+    "ActionResultMessageContent",
     "PromptStack",
     "Reference",
+    "BaseAction",
+    "ToolAction",
 ]