+++ griptape/common/actions/base_action.py
@@ -1,0 +1,5 @@
+from griptape.mixins import SerializableMixin
+from abc import ABC
+
+
+class BaseAction(SerializableMixin, ABC): ...