+++ griptape/common/prompt_stack/messages/delta_message.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
+
 from typing import Optional
 
 from attrs import define, field
 
-from griptape.common.prompt_stack.contents.text_delta_message_content import TextDeltaMessageContent
-
+from griptape.common import BaseDeltaMessageContent
 
 from .base_message import BaseMessage
 
@@ -12,4 +12,4 @@ from .base_message import BaseMessage
 @define
 class DeltaMessage(BaseMessage):
     role: Optional[str] = field(kw_only=True, default=None, metadata={"serializable": True})
-    content: Optional[TextDeltaMessageContent] = field(kw_only=True, default=None, metadata={"serializable": True})
+    content: Optional[BaseDeltaMessageContent] = field(kw_only=True, default=None, metadata={"serializable": True})