+++ griptape/drivers/prompt/openai_chat_prompt_driver.py
@@ -1,28 +1,36 @@
 from __future__ import annotations
 
+import json
 from collections.abc import Iterator
-from typing import Literal, Optional, TYPE_CHECKING
+from typing import TYPE_CHECKING, Literal, Optional
 
 import openai
 from attrs import Factory, define, field
+from schema import Schema
 
-from griptape.artifacts import TextArtifact
+from griptape.artifacts import TextArtifact, ActionArtifact
 from griptape.common import (
+    ActionCallMessageContent,
+    ActionResultMessageContent,
+    BaseDeltaMessageContent,
     BaseMessageContent,
+    ActionCallDeltaMessageContent,
     DeltaMessage,
     TextDeltaMessageContent,
     ImageMessageContent,
     PromptStack,
     Message,
     TextMessageContent,
+    ToolAction,
 )
 from griptape.drivers import BasePromptDriver
 from griptape.tokenizers import BaseTokenizer, OpenAiTokenizer
 
-
 if TYPE_CHECKING:
-    from openai.types.chat.chat_completion_message import ChatCompletionMessage
     from openai.types.chat.chat_completion_chunk import ChoiceDelta
+    from openai.types.chat.chat_completion_message import ChatCompletionMessage
+
+    from griptape.tools import BaseTool
 
 
 @define
@@ -59,6 +67,8 @@ class OpenAiChatPromptDriver(BasePromptDriver):
         default=None, kw_only=True, metadata={"serializable": True}
     )
     seed: Optional[int] = field(default=None, kw_only=True, metadata={"serializable": True})
+    tool_choice: str = field(default="auto", kw_only=True, metadata={"serializable": False})
+    use_native_tools: bool = field(default=True, kw_only=True, metadata={"serializable": True})
     ignored_exception_types: tuple[type[Exception], ...] = field(
         default=Factory(
             lambda: (
@@ -80,8 +90,8 @@ class OpenAiChatPromptDriver(BasePromptDriver):
             message = result.choices[0].message
 
             return Message(
-                content=[self.__message_to_prompt_stack_content(message)],
-                role=message.role,
+                content=self.__to_prompt_stack_message_content(message),
+                role=Message.ASSISTANT_ROLE,
                 usage=Message.Usage(
                     input_tokens=result.usage.prompt_tokens, output_tokens=result.usage.completion_tokens
                 ),
@@ -104,15 +114,9 @@ class OpenAiChatPromptDriver(BasePromptDriver):
                     choice = chunk.choices[0]
                     delta = choice.delta
 
-                    yield DeltaMessage(content=self.__message_delta_to_prompt_stack_content_delta(delta))
+                    yield DeltaMessage(content=self.__to_prompt_stack_delta_message_content(delta))
                 else:
                     raise Exception("Completion with more than one choice is not supported yet.")
-
-    def _prompt_stack_to_messages(self, prompt_stack: PromptStack) -> list[dict]:
-        return [
-            {"role": self.__to_role(message), "content": self.__to_content(message)}
-            for message in prompt_stack.messages
-        ]
 
     def _base_params(self, prompt_stack: PromptStack) -> dict:
         params = {
@@ -120,6 +124,11 @@ class OpenAiChatPromptDriver(BasePromptDriver):
             "temperature": self.temperature,
             "user": self.user,
             "seed": self.seed,
+            **(
+                {"tools": self.__to_openai_tools(prompt_stack.tools), "tool_choice": self.tool_choice}
+                if prompt_stack.tools and self.use_native_tools
+                else {}
+            ),
             **({"stop": self.tokenizer.stop_sequences} if self.tokenizer.stop_sequences else {}),
             **({"max_tokens": self.max_tokens} if self.max_tokens is not None else {}),
             **({"stream_options": {"include_usage": True}} if self.stream else {}),
@@ -130,27 +139,87 @@ class OpenAiChatPromptDriver(BasePromptDriver):
             # JSON mode still requires a system message instructing the LLM to output JSON.
             prompt_stack.add_system_message("Provide your response as a valid JSON object.")
 
-        messages = self._prompt_stack_to_messages(prompt_stack)
+        messages = self.__to_openai_messages(prompt_stack.messages)
 
         params["messages"] = messages
 
         return params
 
-    def __to_role(self, message: Message) -> str:
+    def __to_openai_messages(self, messages: list[Message]) -> list[dict]:
+        openai_messages = []
+
+        for message in messages:
+            if message.is_text():
+                openai_messages.append({"role": message.role, "content": message.to_text()})
+            elif message.has_any_content_type(ActionResultMessageContent):
+                # ToolAction results need to be expanded into separate messages.
+                openai_messages.extend(
+                    [
+                        {
+                            "role": self.__to_openai_role(message),
+                            "content": self.__to_openai_message_content(action_result),
+                            "tool_call_id": action_result.action.tag,
+                        }
+                        for action_result in message.get_content_type(ActionResultMessageContent)
+                    ]
+                )
+            else:
+                # ToolAction calls are attached to the assistant message that originally generated them.
+                action_call_content = []
+                non_action_call_content = []
+                for content in message.content:
+                    if isinstance(content, ActionCallMessageContent):
+                        action_call_content.append(content)
+                    else:
+                        non_action_call_content.append(content)
+
+                openai_messages.append(
+                    {
+                        "role": self.__to_openai_role(message),
+                        "content": [
+                            self.__to_openai_message_content(content)
+                            for content in non_action_call_content  # ToolAction calls do not belong in the content
+                        ],
+                        **(
+                            {
+                                "tool_calls": [
+                                    self.__to_openai_message_content(action_call) for action_call in action_call_content
+                                ]
+                            }
+                            if action_call_content
+                            else {}
+                        ),
+                    }
+                )
+
+        return openai_messages
+
+    def __to_openai_role(self, message: Message) -> str:
         if message.is_system():
             return "system"
         elif message.is_assistant():
             return "assistant"
         else:
-            return "user"
+            if message.has_any_content_type(ActionResultMessageContent):
+                return "tool"
+            else:
+                return "user"
 
-    def __to_content(self, message: Message) -> str | list[dict]:
-        if all(isinstance(content, TextMessageContent) for content in message.content):
-            return message.to_text()
-        else:
-            return [self.__prompt_stack_content_message_content(content) for content in message.content]
+    def __to_openai_tools(self, tools: list[BaseTool]) -> list[dict]:
+        return [
+            {
+                "function": {
+                    "name": f"{tool.name}_{tool.activity_name(activity)}",
+                    "description": tool.activity_description(activity),
+                    "parameters": (tool.activity_schema(activity) or Schema({})).json_schema("Parameters Schema"),
+                },
+                "type": "function",
+            }
+            for tool in tools
+            for activity in tool.activities()
+        ]
 
-    def __prompt_stack_content_message_content(self, content: BaseMessageContent) -> dict:
+    def __to_openai_message_content(self, content: BaseMessageContent) -> str | dict:
         if isinstance(content, TextMessageContent):
             return {"type": "text", "text": content.artifact.to_text()}
         elif isinstance(content, ImageMessageContent):
@@ -158,19 +227,64 @@ class OpenAiChatPromptDriver(BasePromptDriver):
                 "type": "image_url",
                 "image_url": {"url": f"data:{content.artifact.mime_type};base64,{content.artifact.base64}"},
             }
+        elif isinstance(content, ActionCallMessageContent):
+            action = content.artifact.value
+
+            return {
+                "type": "function",
+                "id": action.tag,
+                "function": {"name": action.to_native_tool_name(), "arguments": json.dumps(action.input)},
+            }
+        elif isinstance(content, ActionResultMessageContent):
+            return content.artifact.to_text()
         else:
             raise ValueError(f"Unsupported content type: {type(content)}")
 
-    def __message_to_prompt_stack_content(self, message: ChatCompletionMessage) -> BaseMessageContent:
-        if message.content is not None:
-            return TextMessageContent(TextArtifact(message.content))
-        else:
-            raise ValueError(f"Unsupported message type: {message}")
+    def __to_prompt_stack_message_content(self, response: ChatCompletionMessage) -> list[BaseMessageContent]:
+        content = []
 
-    def __message_delta_to_prompt_stack_content_delta(self, content_delta: ChoiceDelta) -> TextDeltaMessageContent:
-        if content_delta.content is None:
+        if response.content is not None:
+            content.append(TextMessageContent(TextArtifact(response.content)))
+        if response.tool_calls is not None:
+            content.extend(
+                [
+                    ActionCallMessageContent(
+                        ActionArtifact(
+                            ToolAction(
+                                tag=tool_call.id,
+                                name=ToolAction.from_native_tool_name(tool_call.function.name)[0],
+                                path=ToolAction.from_native_tool_name(tool_call.function.name)[1],
+                                input=json.loads(tool_call.function.arguments),
+                            )
+                        )
+                    )
+                    for tool_call in response.tool_calls
+                ]
+            )
+
+        return content
+
+    def __to_prompt_stack_delta_message_content(self, content_delta: ChoiceDelta) -> BaseDeltaMessageContent:
+        if content_delta.content is not None:
+            return TextDeltaMessageContent(content_delta.content)
+        elif content_delta.tool_calls is not None:
+            tool_calls = content_delta.tool_calls
+
+            if len(tool_calls) == 1:
+                tool_call = tool_calls[0]
+                index = tool_call.index
+
+                # Tool call delta either contains the function header or the partial input.
+                if tool_call.id is not None and tool_call.function.name is not None:
+                    return ActionCallDeltaMessageContent(
+                        index=index,
+                        tag=tool_call.id,
+                        name=ToolAction.from_native_tool_name(tool_call.function.name)[0],
+                        path=ToolAction.from_native_tool_name(tool_call.function.name)[1],
+                    )
+                else:
+                    return ActionCallDeltaMessageContent(index=index, partial_input=tool_call.function.arguments)
+            else:
+                raise ValueError(f"Unsupported tool call delta length: {len(tool_calls)}")
+        else:
             return TextDeltaMessageContent("")
-        else:
-            delta_content = content_delta.content
-
-            return TextDeltaMessageContent(delta_content)