+++ tests/unit/common/contents/test_text_delta_message_content.py
@@ -1,0 +1,6 @@
+from griptape.common import TextDeltaMessageContent
+
+
+class TestTextDeltaMessageContent:
+    def test_init(self):
+        assert TextDeltaMessageContent("foo").text == "foo"